pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "v1smartglass"
version = "0.1.0"
description = "Stream Trend Vision One workbench alerts and top risks to Even Realities G2 smart glasses."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [{ name = "V1 SmartGlass G2" }]
keywords = ["trend-micro", "vision-one", "xdr", "even-realities", "smart-glasses", "ble"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
]
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.6",
    "PyYAML>=6.0",
    "bleak>=0.22",
    "Pillow>=10.2",
    "rich>=13.7",
    "typer>=0.12",
]

[project.optional-dependencies]
test = ["pytest>=8.0", "pytest-asyncio>=0.23", "respx>=0.21"]
even = ["even-glasses>=0.3"]

[project.scripts]
v1smartglass = "v1smartglass.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/v1smartglass"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
